% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dawww.R
\name{dawww}
\alias{dawww}
\title{dawww}
\usage{
dawww(choice = c("spring", "summer", "fall", "winter", "smolbabe",
  "doggo", "surprise"))
}
\arguments{
\item{choice}{From the list above, you may only choose one.}
}
\value{
A photo of your choice is downloaded and displayed in the plot window.
}
\description{
Your daily dose of distraction. Make a choice, see the results, smile.
}
\details{
This function is an extension of the original pupR package from https://github.com/melissanjohnson/pupR. All pictures are taken from pixabay.com.
}
\examples{
dawww("smolbabe") #get 1 picture of a smolbabe
replicate(5, "doggo") #get 5 pictures of doggos
}
