% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadSentinel2.R
\name{DownloadSentinel2}
\alias{DownloadSentinel2}
\title{Download Sentinel 2 dataset from google cloud.}
\usage{
DownloadSentinel2(tilename, folder, L2Flag = FALSE, regExp = NULL)
}
\arguments{
\item{tilename}{what tile do you want to download}

\item{folder}{local folder used to store downloaded files}

\item{L2Flag}{indicate to download L2A product or L1C product (default)}

\item{regExp}{regular expression, (default is NULL) with which you can specify which files you want. Be really careful to use this. For example,
regExp <- "S2\\D_MSIL\\d\\D_201([6]|[8][0][0-9]).*.SAFE/$" means you want data from 2016 and before 2018 before October.}
}
\description{
This function requires Google Cloud Storage SDK, you can download and install this SDK
from google cloud website: https://cloud.google.com/sdk/install and make sure the system PATH is set 
so you can use 'gsutil' tool directly in command line.
}
\details{
And just to notice. Most L2A data is only available after 2018, if you want to use L2A before 2018, you need to download L1C data and do the atmophere correction using sen2cor yourself.
You can search if there is L2A data available in your study area using website: https://scihub.copernicus.eu/dhus/#/home , but that website doesn't support batch downloading, and download data 
from there is much more slower than from google cloud.

Once it started, press ctrl+c in the command line can stop it.

If there's any bug, please let me know.

Author: Jacory
}
\examples{
DownloadSentinel2("17SNA", "D:/", L2Flag = TRUE, rexExp = "S2\\\\D_MSIL\\\\d\\\\D_201([6]|[8][0][0-9]).*.SAFE/$")
}
\keyword{sentinel}
