% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_shp_to_spdf.R
\name{url_shp_to_spdf}
\alias{url_shp_to_spdf}
\title{Mass download zipped .shp files from URL}
\usage{
url_shp_to_spdf(URL)
}
\arguments{
\item{URL}{A single URL or a vector of URLs}
}
\value{
A spatialp___dataframe from the desired shapefile is loaded to the global environment.
}
\description{
The function allows you to download shapefile zip files into a temporary folder and directory, which are then unzipped and ready to use in R. Temp folders are automatically unlinked (deleted) and the working directory is set back to original within every iteration of the function. This means no shapefiles are stored to the drive.
}
\details{
Source of function: Kay Cichini (https://www.r-bloggers.com/batch-downloading-zipped-shapefiles-with-r/)
}
\keyword{shapefile}
